requirements.txt
python-telegram-bot==21.6
playwright==1.40.0
selectolax